added search input for every page
