+++ src/app/books/page.tsx
@@ -3,24 +3,13 @@
 import React, { useState } from "react";
 import Link from "next/link";
 import Image from "next/image";
-import { ArrowRight, Search } from "lucide-react";
+import { ArrowRight } from "lucide-react";
 import { books } from "@/data/books";
 import Card from "@/components/Card";
+import SearchInput from "@/components/SearchInput";
 
 export default function Books() {
-  const [searchTerm, setSearchTerm] = useState("");
   const [filteredBooks, setFilteredBooks] = useState(books);
-
-  const handleSearch = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const term = event.target.value.toLowerCase();
-    setSearchTerm(term);
-    const filtered = books.filter(
-      (book) =>
-        book.title.toLowerCase().includes(term) ||
-        book.description.toLowerCase().includes(term),
-    );
-    setFilteredBooks(filtered);
-  };
 
   return (
     <div className="min-h-screen bg-gray-100 dark:bg-black">
@@ -29,21 +18,8 @@ export default function Books() {
           Books to Learn Rust
         </h1>
 
-        <div className="mb-8 max-w-md mx-auto">
-          <div className="relative">
-            <input
-              type="text"
-              placeholder="Search books..."
-              value={searchTerm}
-              onChange={handleSearch}
-              className="w-full p-3 pl-10 rounded-full border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-800 text-gray-900 dark:text-white focus:outline-none focus:ring-2 focus:ring-red-500"
-            />
-            <Search
-              className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400"
-              size={20}
-            />
-          </div>
-        </div>
+        <SearchInput data={books} globalFunc={setFilteredBooks} searchBy={'title,description'}  />
+
 
         <div className="grid md:grid-cols-2 w-full gap-5">
           {filteredBooks.map((book, index) => (
