+++ src/app/devtools/page.tsx
@@ -1,24 +1,12 @@
 "use client";
 
 import React, { useState } from "react";
-import { Search } from "lucide-react";
 import { tools } from "@/data/tools";
 import Card from "@/components/Card";
+import SearchInput from "@/components/SearchInput";
 
 export default function DevTools() {
-  const [searchTerm, setSearchTerm] = useState("");
   const [filteredTools, setFilteredTools] = useState(tools);
-
-  const handleSearch = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const term = event.target.value.toLowerCase();
-    setSearchTerm(term);
-    const filtered = tools.filter(
-      (tool) =>
-        tool.name.toLowerCase().includes(term) ||
-        tool.description.toLowerCase().includes(term),
-    );
-    setFilteredTools(filtered);
-  };
 
   return (
     <div className="min-h-screen bg-gray-100 dark:bg-black">
@@ -27,21 +15,8 @@ export default function DevTools() {
           Rust Developer Tools
         </h1>
 
-        <div className="mb-8 max-w-md mx-auto">
-          <div className="relative">
-            <input
-              type="text"
-              placeholder="Search tools..."
-              value={searchTerm}
-              onChange={handleSearch}
-              className="w-full p-3 pl-10 rounded-full border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-800 text-gray-900 dark:text-white focus:outline-none focus:ring-2 focus:ring-red-500"
-            />
-            <Search
-              className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400"
-              size={20}
-            />
-          </div>
-        </div>
+       <SearchInput data={tools} globalFunc={setFilteredTools} searchBy={'name,description'} />
+
 
         <div className="grid md:grid-cols-3 w-full gap-5">
           {filteredTools.map((tool, index) => (
