+++ src/app/lessons/page.tsx
@@ -2,21 +2,13 @@
 
 import React, { useState } from "react";
 import Link from "next/link";
-import { ArrowRight, Search } from "lucide-react";
+import { ArrowRight } from "lucide-react";
 import { lessons } from "@/data/lessons";
+import SearchInput from "@/components/SearchInput";
 
 export default function Lessons() {
-  const [searchTerm, setSearchTerm] = useState("");
   const [filteredLessons, setFilteredLessons] = useState(lessons);
 
-  const handleSearch = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const term = event.target.value.toLowerCase();
-    setSearchTerm(term);
-    const filtered = lessons.filter((lessons) =>
-      lessons.title.toLowerCase().includes(term),
-    );
-    setFilteredLessons(filtered);
-  };
 
   return (
     <div className="min-h-screen bg-gray-100 dark:bg-gray-900">
@@ -25,21 +17,8 @@ export default function Lessons() {
           Rust Lessons
         </h1>
 
-        <div className="mb-8 max-w-md mx-auto">
-          <div className="relative">
-            <input
-              type="text"
-              placeholder="Search lessons..."
-              value={searchTerm}
-              onChange={handleSearch}
-              className="w-full p-3 pl-10 rounded-full border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-800 text-gray-900 dark:text-white focus:outline-none focus:ring-2 focus:ring-red-500"
-            />
-            <Search
-              className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400"
-              size={20}
-            />
-          </div>
-        </div>
+        <SearchInput data={lessons} globalFunc={setFilteredLessons} searchBy={'title'} />
+
 
         <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
           {filteredLessons.map((lesson, index) => (
